Add a new GH Actions job to automatically update translated document pagse (#598)
This pull request adds a new GitHub Actions job to automate the
translation of document pages.
- Before this job can run, **OPENAI_API_KEY must be added to the project
secrets.**
- It typically takes 8–10 minutes using the o3 model, so the job is
configured to run only when there are changes under docs/ or in
mkdocs.yml.
- The job commits and pushes the translated changes, but it does not
deploy the documents to GitHub Pages. If we think it’s better to deploy
the latest changes automatically as well, I’m happy to update the
workflow. (Personally, I don’t think it’s necessary, since the changes
will be deployed with the next deployment job execution)