Mark some dataclasses as pydantic dataclasses (#1131)
This is the set of top level types used by Temporal for serialization
across activity boundaries. In order to ensure that the models contained
in these dataclasses are built prior to use, the dataclasses need to be
`pydantic.dataclasses.dataclass` rather than `dataclasses.dataclass`

This fixes issues where the types cannot be serialized if the contained
types happen not to have been built. This happens particularly often
when model logging is disabled, which happened to build the pydantic
models as a side effect.